flytt ikonfuknsjonalitet til egen fil
